requirements.txt
absl-py==1.1.0
asgiref==3.5.0
astunparse==1.6.3
autopep8==1.6.0
Babel==2.3.4
bcrypt==3.2.2
beautifulsoup4==4.9.3
cachetools==5.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
decorator==4.0.10
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.11.0
django-filter==21.1
django-passwords==0.3.12
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
docutils==0.12
drf-yasg==1.20.0
ebaysdk==2.1.5
efficientnet==1.0.0
feedparser==5.2.1
flatbuffers==1.12
gast==0.4.0
gevent==1.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.14
grpcio==1.46.3
h5py==3.7.0
html2text==2016.9.19
idna==2.7
image-classifiers==1.0.0
imageio==2.19.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==2.10.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lxml==4.6.2
Mako==1.0.4
Markdown==3.3.6
MarkupSafe==0.23
matplotlib==3.3.4
mock==2.0.0
mysql-connector-python==8.0.17
networkx==2.6.3
num2words==0.5.6
numpy==1.21.6
oauthlib==3.2.0
ofxparse==0.16
openapi-codec==1.3.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
passlib==1.6.5
pbr==5.5.1
phonenumbers==8.12.16
Pillow==9.1.1
pipreqs==0.4.11
protobuf==3.19.4
psutil==5.6.3
psycopg2==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydot==1.2.3
PyJWT==2.4.0
pymongo==3.12.3
PyOpenGL @ file:///C:/Users/David%20Speedster.000/Downloads/PyOpenGL-3.1.5-cp37-cp37m-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/David%20Speedster.000/Downloads/PyOpenGL_accelerate-3.1.5-cp37-cp37m-win_amd64.whl
pyparsing==2.1.10
PyPDF2==1.26.0
pyserial==3.1.1
python-dateutil==2.5.3
python-stdnum==1.14
python3-openid==3.2.0
pytz==2016.7
pyusb==1.0.0
PyWavelets==1.3.0
pywin32==300
PyYAML==3.13
qrcode==5.3
reportlab==3.5.60
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scipy==1.7.3
segmentation-models==1.0.1
simplejson==3.17.6
six==1.15.0
soupsieve==2.1
sqlparse==0.2.4
suds-jurko==0.6
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
tqdm==4.64.0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.3
Werkzeug==2.1.2
wrapt==1.14.1
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
yarg==0.1.9
zipp==3.8.0
